requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
mysql-connector-python==8.3.0
email-validator==2.1.0.post1 